Copy all local packages required by `@expo/worker`

+++ packages/worker/package.sh
@@ -33,8 +33,12 @@ copy_package() {
   dst=$2
   mkdir -p $dst
   cp "$src/package.json" "$dst/package.json"
-  cp "$src/tsconfig.json" "$dst/tsconfig.json"
-  cp "$src/tsconfig.build.json" "$dst/tsconfig.build.json"
+  if [[ -f "$src/tsconfig.json" ]]; then
+    cp "$src/tsconfig.json" "$dst/tsconfig.json"
+  fi
+  if [[ -f "$src/tsconfig.build.json" ]]; then
+    cp "$src/tsconfig.build.json" "$dst/tsconfig.build.json"
+  fi
   cp -r "$src/src" "$dst/src"
 }
 
@@ -43,11 +47,16 @@ cp "$ROOT_DIR/package.json" $target_root_dir
 cp "$ROOT_DIR/lerna.json" $target_root_dir
 
 copy_package $WORKER_DIR $target_worker_dir
+while IFS= read -r package_dir; do
+  package_name=$(basename "$package_dir")
+  copy_package "$package_dir" "$target_root_dir/packages/$package_name"
+done <<< "$(yarn --silent lerna list --scope "@expo/worker" --include-dependencies --parseable --loglevel silent)"
+
 
 pushd $target_root_dir >/dev/null 2>&1
-yarn install --silent
+yarn install --silent --frozen-lockfile
 yarn build
-yarn install --silent --production=true
+yarn install --silent --production=true --frozen-lockfile
 rm -rf tsconfig.json tsconfig.build.json
 popd >/dev/null 2>&1
